# Less allocatey ethash