--- a/src/Nethermind/Nethermind.HashLib/Crypto/SHA3/KeccakBase.cs
+++ b/src/Nethermind/Nethermind.HashLib/Crypto/SHA3/KeccakBase.cs
@@ -77,7 +77,7 @@ namespace Nethermind.HashLib.Crypto.SHA3
             // TODO: review discussions on whether it is always safe (should be)
             ReadOnlySpan<ulong> data = MemoryMarshal.Cast<byte, ulong>(a_data.Slice(a_index, _blockSize));
             //ulong[] data= Converters.ConvertBytesToULongs(a_data, a_index, BlockSize);
-            
+
             for (int j = 0; j < _blockSize / 8; j++)
                 m_state[j] ^= data[j];
 
@@ -150,11 +150,11 @@ namespace Nethermind.HashLib.Crypto.SHA3
             Aso = m_state[23];
             Asu = m_state[24];
 
-            Ca = Aba^Aga^Aka^Ama^Asa;
-            Ce = Abe^Age^Ake^Ame^Ase;
-            Ci = Abi^Agi^Aki^Ami^Asi;
-            Co = Abo^Ago^Ako^Amo^Aso;
-            Cu = Abu^Agu^Aku^Amu^Asu;
+            Ca = Aba ^ Aga ^ Aka ^ Ama ^ Asa;
+            Ce = Abe ^ Age ^ Ake ^ Ame ^ Ase;
+            Ci = Abi ^ Agi ^ Aki ^ Ami ^ Asi;
+            Co = Abo ^ Ago ^ Ako ^ Amo ^ Aso;
+            Cu = Abu ^ Agu ^ Aku ^ Amu ^ Asu;
             Da = Cu ^ (Ce << 1) ^ (Ce >> (64 - 1));
             De = Ca ^ (Ci << 1) ^ (Ci >> (64 - 1));
             Di = Ce ^ (Co << 1) ^ (Co >> (64 - 1));
@@ -2724,6 +2724,11 @@ namespace Nethermind.HashLib.Crypto.SHA3
             Buffer.BlockCopy(m_state, 0, result, 0, HashSize);
             return result;
             //return Converters.ConvertULongsToBytes(m_state).SubArray(0, HashSize);
+        }
+
+        protected override void GetResultUints(Span<uint> result)
+        {
+            m_state.AsSpan(0, HashSize / sizeof(ulong)).CopyTo(MemoryMarshal.Cast<uint, ulong>(result));
         }
 
         protected override uint[] GetResultUInts()