--- a/src/Nethermind/Nethermind.HashLib/Nethermind.HashLib.csproj
+++ b/src/Nethermind/Nethermind.HashLib/Nethermind.HashLib.csproj
@@ -7,6 +7,7 @@
     <Compile Remove="Hash32.cs" />
   </ItemGroup>
   <ItemGroup>
+    <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="4.6.0" />
     <PackageReference Include="System.Security.Cryptography.Algorithms" Version="4.3.1" />
   </ItemGroup>
 </Project>