--- a/src/Nethermind/Nethermind.Mining/IEthashDataSet.cs
+++ b/src/Nethermind/Nethermind.Mining/IEthashDataSet.cs
@@ -21,6 +21,6 @@ namespace Nethermind.Mining
     public interface IEthashDataSet : IDisposable
     {
         uint Size { get; }
-        uint[] CalcDataSetItem(uint i);
+        void CalcDataSetItem(uint i, Span<uint> output);
     }
 }