An evening play with ethash resulting in the following improvement showed by benchmark

|   Method |     Mean |     Error |   StdDev |Ratio | Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |---------:|----------:|---------:|-----:|------:|------:|------:|----------:|
| Before   | 25.07 ms | 0.6110 ms | 1.733 ms | 1.00 |     - |     - |     - | 261.91 KB |
| After changing Hashimoto   | 24.04 ms | 0.4697 ms | 0.8707 ms| 1.00 |     - |     - |     - | 223.91 KB |
| After improving EthashCache | 24.53 ms | 0.4897 ms | 0.8831 ms |  1.00 |     - |     - |     - |   2.84 KB |


The first important improvement was moving the allocation of the array in `Hashimoto` outside of the loop. The second, was improving `Keccak512` to be able to run it on `Span<uint>`, which caused the benchmark to drop allocation 100x.